Change BaseAlgorithm to now inherit Concept
