+++ src/orion/core/worker/experiment.py
@@ -16,6 +16,7 @@ import logging
 import random
 import sys
 
+from orion.algo.base import PrimaryAlgo
 from orion.core.cli.evc import fetch_branching_configuration
 from orion.core.evc.adapters import Adapter, BaseAdapter
 from orion.core.evc.conflicts import detect_conflicts
@@ -24,7 +25,6 @@ from orion.core.io.experiment_branch_builder import ExperimentBranchBuilder
 from orion.core.io.interactive_commands.branching_prompt import BranchingPrompt
 from orion.core.io.space_builder import SpaceBuilder
 from orion.core.utils.format_trials import trial_to_tuple
-from orion.core.worker.primary_algo import PrimaryAlgo
 from orion.core.worker.trial import Trial
 
 log = logging.getLogger(__name__)
